Move generic properties from VehicleStatus to Vehicle
